Introduce basic TheTrainline API connection
All the Trainline API requests should be based on Faraday for
some reasons. The Trainline connection is expected to
handle the requests in a fast way. So the Faraday object should
be cached in order to avoid the same request setup across multiple
requests. The connection should benefit from a dedicated middleware
that handles all the request boilerplate like timeout,
JSON-parsing or exception handling.