Request locations from TheTrainline API
We investigated TheTrainline API by doing some
research on their official website and what requests
are performed with the browser client. Turned out
several requests to different API endpoints are
performed. First and foremost the locations for
departure and arrival are requested for their
location-code. That location-code is used for
searching the journeys. So before implementing
the journey search itself we should prototype the
required location requests.