Consider zipped responses
Turns out the Trainline journey search response
is pretty verbose.
Some measurements exposed > 300K payload on average.
So we shhould consider zipping the response via the net
in order to speed up latencies.
For that we have to enrich the connection to request
zipped content on demand and enable the middleware to
unzip it. So we go with some dependency injection by
introducing and passing the ZippedJasonParser.